move URLSession static methods to utility
