+++ Tests/ParseSwiftTests/ExtensionsTests.swift
@@ -49,7 +49,7 @@ class ExtensionsTests: XCTestCase {
 
     func testReconnectInterval() throws {
         for index in 1 ..< 50 {
-            let time = URLSession.reconnectInterval(index)
+            let time = Utility.reconnectInterval(index)
             XCTAssertLessThan(time, 30)
             XCTAssertGreaterThan(time, -1)
         }
